integration tests extracted
